feat: add burn/claim page (#122)
* feat: add burn/claim page

this first commit is a draft for the ongoing work on this page. for starters, i've updated the useTheme hook to be type-safe.

i also went on to create a global store to sync the wallet state, although this is just a PoC for the entire implementation, using local state proved inefficient, as the click of "connect wallet" on the main card did not update the button state in the TopBar component.

i believe that this sets up a ground work for how we can manage reactivity and sync UI state accorss the application with zustand.

the UI still needs to match what's on the design, which i'll work on, and also finalize the process of adding new design tokens (color codes, basically) from figma, so i can properly implement the dark/light theme styles.

* feat(ui): a complete refactor of the page to match what's on Figma

i've also included more components to proper reflect these changes.

the "connect wallet" button is now extensible with a new Gradient component to mimick the gradient border we have on most of the UI components on Figma. i also included new design tokens for the dark mode.

moving on to the "light mode" implementation now.

* feat: add light theme styles and refactor the theme hook to use a context provider

the previous approach of enabling theming in the UI was buggy as there were stale values, and the body of the page doesn't recieve the correct theme value when it is switched.

each component tries to call it on its own, which led to race conditions. when you switch to dark mode, some components' style would still retain light mode style (and vice versa, unless you do a "hard-reload"), which is problematic.

the solution here was to move the logic into a global store with React context, so all components, irrespective of where they are in the tree have access to the same theme context.