feat: add dashboard UI (#129)
* fix: fixed mismatched imports due to the recent PR

i've also updated the ThemeProvider to expose a isDark value so it is not computed everytime in components needing it.

* feat: add dashboard UI

this is a work-in-progress for the dashboard UI.

for now, i have teh cards and the charts hooked-up onto dummy data. next stop is refactoring to match the designs acurately and fixing the modal UI for connecting a new wallet.

* feat: add "connect wallet" modal. a tabbed-chart UI and realtime asset data

this commit included a refactor of the dashboard to macth the UI design on figma, the connected wallet state, "connect wallet" modal, ensuring that the dashbaord content doesn't stretch on extra-large screens, and pulling asset data from a live DEX API

* feat: add ability to connect to a real wallet

this commit introduces the feature that lets people connect their real wallet accoun from the Ethereum or Starknet network with support for various wallets

* feat: add dynamic wallet provider detection to the wallet-connection flow

we still need to circle back on the idea of listing the available starknet providers when people have more than one starknet wallet installed.

should be a rare case. but we shall consider it.

* fix: show a list of starknet wallet providers when people have more than one extension installed

* render modal in the app layout